Update genesis.py
Line 1456 tests for beam.len() > 1 and then assigns inp.beam. the case beam.len() !>1 is not treated which will then cause an exception in line 1491. This commit wraps lines 1496 in a try/except statement and tries to remedy the case beam.len() !> 1. No functionality nor regression tests were performed to justify this fix, so needs more testing and review.