Some nice new things for ParticleArray class
a few better comments

gamma property and beta property for getting these relativistic quantities for each
macroparticle.

sort method for sorting the array by either one of the main callable names (x, px, y,
etc) or one of the properties, like gamma, beta, and so on.

implement __len__ method for returning the number of particles in the array.

restore get_envelope to calculate also the normalised emittances, not just the geometric ones.