Options to return pd.DataFrame instead of list of Twiss instances
an iterable of Twiss instances is difficult and unwieldy.  Having to do list
comprehensions to look at any variable is annoying.  To manipulate the data will then
also require a conversion to a np.array.  Far easier to just use a Pandas DataFrame.  I
also include a method for converting a Twiss instance to a Pandas Series.  The kwarg
return_df are for the two functions track and twiss.