Fix bug caused by having matplotlib as a CSR instance attribute
matplotlib.pyplot is a module.  Modules cannot be picked.  This means instances of
the class cannot be deepcopied (or for example used in multiprocessing...), so best
really to not do this.  This leads to crashes in for example
Navigator.add_physics_proc when you have added CSR with pict_debug enabled to the
Navigator.  In essence this meakes pict_debug unusable.