astra2ocelot, charge cannot be negative in ocelot, -ve implied.
If for some reason the charge being read from OCELOT is positive, then you get a
negative charge when multiplying by -1.  this gives nonsense results, positive
charge is assumed in a number of places in ocelot.  indeed in the astra manual it
says:

"The sign of the charge specified in the column 8 [macro charge] is not relevant".